Native mutation support - Incremental PR - I (#118)
https://hasurahq.atlassian.net/browse/ES-99

<!-- The PR description should answer 2 (maybe 3) important questions:
-->

### What

<!-- What is this PR trying to accomplish (and why, if it's not
obvious)? -->

This PR does the following things:

1. Parses `nativeMutations` from the configuration.json file.
2. Generates the NDC schema of the native mutations.

### How

<!-- How is it trying to accomplish it (what are the implementation
steps)? -->

The schema of a native mutation is created in the following way:

1. A native mutation creates two object types:
a. Object that corresponds to the `columns` that will capture the output
of the native query.
b. Object that will be the response of the native query with
`affected_rows` and `returning` being the fields of the object.