Native mutations execution incremental PR (Incremental PR - III) (#131)
<!-- The PR description should answer 2 (maybe 3) important questions:
-->

### What

<!-- What is this PR trying to accomplish (and why, if it's not
obvious)? -->

This PR is the third incremental PR that follows
https://github.com/hasura/ndc-sqlserver/pull/118 and
https://github.com/hasura/ndc-sqlserver/pull/130. This PR adds the
execution logic of the native mutations. Please note that, tests will be
added in a separate PR because this PR as it is crosses 1000 lines diff,
but most of these changes are cosmetic changes and the core logic of
this PR should be around 500-600 LOC.



### How

<!-- How is it trying to accomplish it (what are the implementation
steps)? -->

This PR adds the execution logic of the native mutations in the
following manner:

1. Parse a `MutationRequest` and then convert it into a
`MutationExecutionPlan`.
2. The `MutationExecutionPlan` contains a `NativeMutationExecutionPlan`
containing relevant information for a native mutation to be executed.
3. The native mutation query is run first and then the rows obtained
from the database are converted into a JSON array of objects, where each
object represents a single row.
4. A second query is made per mutation operation, where the JSON from
step 3 is provided as a parameter and the fields requested in the NDC
query are returned in the appropriate format.