read connection string from environment variable (#122)
<!-- The PR description should answer 2 (maybe 3) important questions:
-->

### What

The PR add the ability to use an Env Var to read the MSSQL connection
string, instead of hard-coding the value in the config file (now the
config file can contain the name of the env var, which can be provided
while starting the server)

By default, the CLI will add the env var `CONNECTION_URI` for the
string.

<!-- What is this PR trying to accomplish (and why, if it's not
obvious)? -->

### How

<!-- How is it trying to accomplish it (what are the implementation
steps)? -->