Stored procedures schema and introspection (Incremental PR - I) (#142)
<!-- The PR description should answer 2 (maybe 3) important questions:
-->

### What

<!-- What is this PR trying to accomplish (and why, if it's not
obvious)? -->

This PR adds introspection and schema support for stored procedures in
MSSQL. This PR is the first incremental PR towards adding stored
procedures support.

### How

<!-- How is it trying to accomplish it (what are the implementation
steps)? -->

### Introspection

A database query is made to get all the metadata about stored procedures
present in the database. As it is a stored procedure, there is no return
type tied to it, so we leave the `returns` part empty and we expect the
user to fill in the appropriate schema.

### Schema

We do schema generation for a stored procedure, only if it has a
`returns` associated with it. Otherwise, we won't be able to generate a
graphql schema for the stored procedure.