Stored procedures execution (Incremental PR - II) (#141)
<!-- The PR description should answer 2 (maybe 3) important questions:
-->

### What

<!-- What is this PR trying to accomplish (and why, if it's not
obvious)? -->

This PR is the second incremental PR towards adding stored procedures
support and this PR adds the execution logic of the same. The actual PR
logic is not very large but due to the tests, the lines diff is around
18k.

### How

<!-- How is it trying to accomplish it (what are the implementation
steps)? -->

Steps:

1. Identify whether the procedure name in the mutation request is a
stored procedure.
2. Fetch the metadata of the stored procedure.
3. Arguments and field validation, throw error if any required argument
is not provided in the request.
4. Create a temporary table, the schema of the temporary table is
determined by the `returns` field of the stored procedure. The temporary
table is used to write the results of the stored procedure into.
5. The SQL query is generated in the following format:

```sql
INSERT INTO #TempTable (col1, col2) EXEC <stored_procedure_name> @arg1_name = arg_value 
```

6. Make another SQL query to query from the temporary table with the
fields requested and return the response.