Add a subcommand to introspect stored procedures (#148)
<!-- The PR description should answer 2 (maybe 3) important questions:
-->

### What

<!-- What is this PR trying to accomplish (and why, if it's not
obvious)? -->

This PR adds a subcommand `stored-procedures` to the CLI `update`
command to introspect stored procedures separately. This is because the
stored procedures introspection doesn't include any return type and a
stored procedure without any return type is not added to the connector's
schema. After the introspection, the user needs to manually add the
return types of the stored procedures and this information should be
retained, that's why the introspection of stored procedures is decoupled
from the introspection of the `tables` , `types` etc.

### How

<!-- How is it trying to accomplish it (what are the implementation
steps)? -->